Computer Simulation project
This project includes the files: Hex.java, TasmanianDevil.java, World.java. The directions for the project where as follows: 
Write classes to complete a simulation of devils moving around a hex-grid world infecting each other with devil facial tumor disease (DFTD). A world will be specified by the width of the grid (the height should be chosen so that the world is roughly square in terms of Cartesian coordinates), the total number of devils in it, and the number of devils initially infected with DFTD. Each devil moves around the world by randomly choosing a location on the grid and moving one hex closer to its target during each step of the simulation. When a devil reaches its goal, it randomly chooses another goal. When two devils are at the same location and one of them is infected with DFTD and the other is not, there is an 80% chance that the other devil becomes infected (if more than two devils are at the same location, this interaction should be repeated for each pair of d